Combine everything into one file
